requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-extensions==2.0.0
django-filter==1.1.0
django-js-asset==1.0.0
django-mptt==0.9.0
djangorestframework==3.7.7
idna==2.6
Markdown==2.6.11
ordering==0.1.0
Pygments==2.2.0
pytz==2018.3
requests==2.18.4
six==1.11.0
typing==3.6.4
urllib3==1.22
